Issue #8: Added support for Throwables
The Try API was adapted to catch Throwables but to rethrow fatal Errors.
In this way, the Try type adheres to the pattern of e.g. Optional and
CompleteableFuture and is, thus, easier to integrate. The
FailableSupplier was changed to throw Throwables instead of Exceptions.
Further, the corresponding method signatures in the Try type were
changed from Exception to Throwable. Unit test is also added.